삭제하는 중
오류가 계속 떠서 물어볼 예정

+++ daily_health_server/src/main/java/com/example/daily_health_server/controller/FoodsRecordController.java
@@ -24,7 +24,7 @@ public class FoodsRecordController {
 		food.setFoodId(anfood.getFoodId());
 		
 		User user = new User();
-		user.setUserid(Long.valueOf(anfood.getUserid()));
+		user.setUserid(Long.valueOf(anfood.getUserid()).intValue());
 		foodRecord.setFood(food);
 		foodRecord.setMonth(anfood.getMonth());
 		foodRecord.setDay(anfood.getDay());
